services/news-monitor/requirements.txt
feedparser>=6.0
httpx>=0.27
openai>=1.0
python-telegram-bot>=21.0
pyyaml>=6.0
psycopg2-binary>=2.9
beautifulsoup4>=4.12
